TestCase: remove the generic stubbing of the translation and escaping functions
Only a very limited number of test need these stubs, so we may as well only load them if and when needed.

Includes removing some stubbing being done in certain test functions for a limited set of these function, which were a) previously already covered in the test `setUp()` and b) are better served by using the dedicated functions for stubbing the complete set of functions.