Tests: switch over to the Yoast\WPTestUtils\BrainMonkey\TestCase for one specific test class
Three tests within the `Schema_Test` test class set expectations for the `wp_strip_all_tags()` method.

However, the `wp_strip_all_tags()` method is now stubbed by default in the `YoastTestCase` `set_up()` method, which in effects means that the calls to the `wp_strip_all_tags()` method are not seen by Mockery and the expectations would not be met.

To maintain the existing test behaviour:
* The parent `TestCase` for this test class is switched from the `Yoast\WP\Woocommerce\Tests\TestCase` to the `Yoast\WPTestUtils\BrainMonkey\TestCase` which doesn't contain the stubbing of the `wp_strip_all_tags()` method;
* The expectations for the `get_option*()` methods are copied from the `Yoast\WP\Woocommerce\Tests\TestCase` `set_up()` method to the `set_up()` method in this test class.class
* And the reason for using a different parent `TestCase` are documented in the `set_up()` method docblock.