WPSEO_WooCommerce_UnitTestCase: switch parent class
This switches the parent class of the `WPSEO_WooCommerce_UnitTestCase` to the WP Test Utils `Yoast\WPTestUtils\WPIntegration\TestCase`.

Includes:
* Removing the `expectOutput()` method.
    This method is not used in this test suite and shouldn't be needed anyway as PHPUnit has build-in `expectOutputString()` and `expectOutputRegex()` methods which should be used instead.
    Additionally, WP Test Utils contains an `expectOutputContains()` method to complement the PHPUnit native methods.
* Removing other methods which were unused in this test suite.
    This applies to all the other methods.

This leaves the `WPSEO_WooCommerce_UnitTestCase` as an opaque class.
I've elected to let it remain though for the following reasons:
* Allows for adding future _plugin-specific_ test helper methods.
* Convenience as all test cases for this plugin `extend` from this one and any future changes in inheritance would only need to be applied to this class to apply them effectively to all test classes for this plugin.

Lastly, this commit makes sure that the generic test case is declared as `abstract` as it doesn't contain any tests itself.