QA: use the most specific PHPUnit assertion possible
This is a long established best practice.

PHPUnit contains a variety of assertions and the ones available have been extended hugely over the years.

To have the most reliable tests, the most specific assertion should be used.

This implements this for this plugin.

Most notably, this changes calls to `assertEquals()` to `assertSame()`, where `assertEquals()` does a loose type comparison and `assertSame()` does a strict type comparison.

Refs:
* https://phpunit.de/manual/6.5/en/appendixes.assertions.html#appendixes.assertions.assertEquals
* https://phpunit.de/manual/6.5/en/appendixes.assertions.html#appendixes.assertions.assertSame