Travis: update for using correct PHPUnit version with WP 5.9+
As of WP 5.9, the integration tests can run on PHPUnit 5.7 - 9.x and should use the most appropriate PHPUnit version for the PHP version on which the tests are being run.

For WP < 5.9, the range of supported PHPUnit versions is still PHPUnit 5.7 - 7.x.

With that in mind and with there being a) a `config - platform - php` setting and b) a committed `composer.lock` file, some version juggling needs to be done.

For WP < 5.9, the logic previously used remains in place.
For WP 5.9, new logic is introduced.

As the conditions if they were to be made stand-alone would get pretty complicated, I've resorted to using `if - elif` conditions instead of stand-alone conditions.

And to make sure the various composer changes don't interfere with each other between the test runs for the integration tests and the Brainmonkey tests, a git reset of the `composer.lock` and the `composer.json` files and a `composer install` is run between the integration tests and the unit tests runs.