Set `Update URI` header in the plugin file.
To utilize the changes from https://make.wordpress.org/core/2021/06/29/introducing-update-uri-plugin-header-in-wordpress-5-8/ .

Note that I did not use a shortlink. This is because the host of the mentioned URI can be (but is not yet) used in the code later on and I don't think we want that to be `yoa.st`.