GH Actions/integration test: test against MySQL 5.6
This syncs the logic for determining the MySQL version to use in the integration tests with the logic used by the majority of plugins.

The difference is historic as the original script was based on what was being done in Travis, which was not always the same between plugins.

This commit is part of a standardization round.